Procfile, restarting
I had trouble with the services restarting so I decided to try to use
foreman setup the environment.  I couldn't get foreman installed but
that lead me to just using the foreman library to generate supervisord
configuration based on the environment and the Procfile.  This worked
well.

I also demonstrated how to load a different environment based the chef
environment.  This should be everything we need to get started.

I will like to extend this a bit and add support for static files but
for now I'm pretty happy with it.